+++ packages/plugins/typescript/react-apollo-offix/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @graphql-codegen/typescript-react-offix
 
+## 1.17.13
+
+### Patch Changes
+
+- d9212aa0: fix(visitor-plugin-common): guard for a runtime type error
+- Updated dependencies [d9212aa0]
+- Updated dependencies [f0b5ea53]
+- Updated dependencies [097bea2f]
+  - @graphql-codegen/visitor-plugin-common@1.20.0
+  - @graphql-codegen/plugin-helpers@1.18.5
+
 ## 1.17.12
 
 ### Patch Changes